eclipse/rdf4j#998 simplify tests
