+++ shacl/src/test/java/org/eclipse/rdf4j/sail/shacl/ShaclSailSupportedPredicatesDocumentationTest.java
@@ -1,0 +1,78 @@
+/*******************************************************************************
+ * Copyright (c) 2018 Eclipse RDF4J contributors.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Distribution License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/org/documents/edl-v10.php.
+ *******************************************************************************/
+
+package org.eclipse.rdf4j.sail.shacl;
+
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Model;
+import org.eclipse.rdf4j.model.vocabulary.SHACL;
+import org.eclipse.rdf4j.rio.RDFFormat;
+import org.eclipse.rdf4j.rio.Rio;
+import org.junit.AfterClass;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import static junit.framework.TestCase.assertTrue;
+
+/**
+ * @author Håvard Ottestad
+ */
+@RunWith(Parameterized.class)
+public class ShaclSailSupportedPredicatesDocumentationTest extends AbstractShaclTest {
+
+
+	private static HashSet<IRI> staticShaclPredicates = new HashSet<>(ShaclSail.getSupportedShaclPredicates());
+
+	public ShaclSailSupportedPredicatesDocumentationTest(String testCasePath, String path, ExpectedResult expectedResult) {
+		super(testCasePath, path, expectedResult);
+	}
+
+
+	@AfterClass
+	public static void afterClass() {
+
+		assertTrue("No test uses the following predicate that the ShaclSail announces as supported: " + Arrays.toString(staticShaclPredicates.toArray()), staticShaclPredicates.isEmpty());
+	}
+
+
+	@Test
+	public void testShaclSailSupportedPredicatesDocumentation() throws IOException {
+
+		HashSet<IRI> shaclPredicates = new HashSet<>(ShaclSail.getSupportedShaclPredicates());
+
+		Model parse = getShacl();
+
+
+		Set<IRI> predicatesInUseInTest = parse.predicates().stream().filter(p -> p.getNamespace().equals(SHACL.NAMESPACE)).collect(Collectors.toSet());
+
+		for (IRI predicate : predicatesInUseInTest) {
+			assertTrue("Predicate used in test but not listed in ShaclSail: " + predicate, shaclPredicates.contains(predicate));
+			staticShaclPredicates.remove(predicate);
+		}
+
+
+	}
+
+
+	private Model getShacl() throws IOException {
+		String shaclFile = getShaclPath();
+
+		return Rio.parse(ShaclSailSupportedPredicatesDocumentationTest.class.getClassLoader().getResourceAsStream(shaclFile), "", RDFFormat.TURTLE);
+	}
+
+
+
+
+}
