+++ shacl/src/test/java/org/eclipse/rdf4j/sail/shacl/ShaclTest.java
@@ -44,57 +44,13 @@ import static org.junit.Assert.assertFalse;
  * @author Håvard Ottestad
  */
 @RunWith(Parameterized.class)
-public class ShaclTest {
+public class ShaclTest extends AbstractShaclTest{
 
-	private static final List<String> testCasePaths = Arrays.asList(
-		"test-cases/datatype/simple",
-		"test-cases/minLength/simple",
-		"test-cases/maxLength/simple",
-		"test-cases/pattern/simple",
-		"test-cases/languageIn/simple",
-		"test-cases/nodeKind/simple",
-		"test-cases/minCount/simple",
-		"test-cases/maxCount/simple",
-		"test-cases/or/inheritance",
-		"test-cases/or/inheritance-deep",
-		"test-cases/or/inheritance-deep-minCountMaxCount",
-		"test-cases/or/inheritanceNodeShape",
-		"test-cases/or/datatype",
-		"test-cases/or/minCountMaxCount",
-		"test-cases/or/maxCount",
-		"test-cases/or/minCount",
-		"test-cases/or/nodeKindMinLength",
-		"test-cases/or/implicitAnd",
-		"test-cases/or/datatypeDifferentPaths"
 
-	);
-
-	private final String testCasePath;
-	private final String path;
-	private final ExpectedResult expectedResult;
-
-	public ShaclTest(String testCasePath, String path, ExpectedResult expectedResult) {
-		this.testCasePath = testCasePath;
-		this.path = path;
-		this.expectedResult = expectedResult;
-		LoggingNode.loggingEnabled = true;
+	public ShaclTest(String testCasePath, String path, AbstractShaclTest.ExpectedResult expectedResult) {
+		super(testCasePath, path, expectedResult);
 	}
 
-	@Parameterized.Parameters(name = "{2} - {1}")
-	public static Collection<Object[]> data() {
-
-		return getTestsToRun();
-	}
-
-	private static HashSet<IRI> staticShaclPredicates = new HashSet<>(ShaclSail.getSupportedShaclPredicates());
-
-
-	@AfterClass
-	public static void afterClass(){
-
-
-		assertTrue("No test uses the following predicate that the ShaclSail announces as supported: "+Arrays.toString(staticShaclPredicates.toArray()), staticShaclPredicates.isEmpty());
-	}
 
 	@Test
 	public void test() throws Exception {
@@ -104,207 +60,6 @@ public class ShaclTest {
 	@Test
 	public void testSingleTransaction() throws Exception {
 		runTestCaseSingleTransaction(testCasePath, path, expectedResult);
-	}
-
-	@Test
-	public void testShaclSailSupportedPredicatesDocumentation() throws IOException {
-
-		HashSet<IRI> shaclPredicates = new HashSet<>(ShaclSail.getSupportedShaclPredicates());
-
-		Model parse = getShacl();
-
-
-		Set<IRI> predicatesInUseInTest = parse.predicates().stream().filter(p -> p.getNamespace().equals(SHACL.NAMESPACE)).collect(Collectors.toSet());
-
-		for (IRI predicate : predicatesInUseInTest) {
-			assertTrue("Predicate used in test but not listed in ShaclSail: "+predicate, shaclPredicates.contains(predicate));
-			staticShaclPredicates.remove(predicate);
-		}
-
-
-	}
-
-	private Model getShacl() throws IOException {
-		String shaclPath = testCasePath;
-
-		if (!shaclPath.endsWith("/")) {
-			shaclPath = shaclPath + "/";
-		}
-
-		String shaclFile = shaclPath + "shacl.ttl";
-
-		return Rio.parse(ShaclTest.class.getClassLoader().getResourceAsStream(shaclFile), "", RDFFormat.TURTLE);
-	}
-
-	private static List<String> findTestCases(String testCase, String baseCase) {
-
-		List<String> ret = new ArrayList<>();
-
-		for (int i = 0; i < 1000; i++) {
-			String path = testCase + "/" + baseCase + "/case" + i;
-			InputStream resourceAsStream = ShaclTest.class.getClassLoader().getResourceAsStream(path);
-			if (resourceAsStream != null) {
-				ret.add(path);
-				try {
-					resourceAsStream.close();
-				}
-				catch (IOException e) {
-					throw new RuntimeException(e);
-				}
-			}
-		}
-
-		return ret;
-
-	}
-
-	private static Collection<Object[]> getTestsToRun() {
-		List<Object[]> ret = new ArrayList<>();
-
-		for (String testCasePath : testCasePaths) {
-			for (ExpectedResult baseCase : ExpectedResult.values()) {
-				findTestCases(testCasePath, baseCase.name()).forEach(path -> {
-					Object[] temp = { testCasePath, path, baseCase };
-					ret.add(temp);
-
-				});
-			}
-		}
-
-		return ret;
-	}
-
-	private static void runTestCase(String shaclPath, String dataPath, ExpectedResult expectedResult) throws Exception {
-
-		if (!dataPath.endsWith("/")) {
-			dataPath = dataPath + "/";
-		}
-
-		if (!shaclPath.endsWith("/")) {
-			shaclPath = shaclPath + "/";
-		}
-
-		String shaclFile = shaclPath + "shacl.ttl";
-		System.out.println(shaclFile);
-		ShaclSail shaclSail = new ShaclSail(new MemoryStore());
-		shaclSail.setLogValidationPlans(true);
-		SailRepository shaclRepository = new SailRepository(shaclSail);
-		shaclRepository.initialize();
-		Utils.loadShapeData(shaclRepository, shaclFile);
-
-		boolean exception = false;
-		boolean ran = false;
-
-		for (int j = 0; j < 100; j++) {
-
-			String name = dataPath + "query" + j + ".rq";
-			try (InputStream resourceAsStream = ShaclTest.class.getClassLoader().getResourceAsStream(name)) {
-				if (resourceAsStream == null) {
-					continue;
-				}
-
-				ran = true;
-				System.out.println(name);
-
-				try (SailRepositoryConnection connection = shaclRepository.getConnection()) {
-					connection.begin(IsolationLevels.SNAPSHOT);
-					String query = IOUtil.readString(resourceAsStream);
-					connection.prepareUpdate(query).execute();
-					connection.commit();
-				} catch (RepositoryException sailException) {
-					exception = true;
-					System.out.println(sailException.getMessage());
-
-				System.out.println("\n############################################");
-				System.out.println("\tValidation Report\n");
-				ShaclSailValidationException cause = (ShaclSailValidationException) sailException.getCause();
-				Model validationReport = cause.validationReportAsModel();
-				Rio.write(validationReport, System.out, RDFFormat.TURTLE);
-				System.out.println("\n############################################");
-				}
-			} catch (IOException e) {
-				e.printStackTrace();
-			}
-
-		}
-
-		shaclSail.shutDown();
-
-		if (ran) {
-			if (expectedResult == ExpectedResult.valid) {
-				assertFalse("Expected transaction to succeed", exception);
-			} else {
-				assertTrue("Expected transaction to fail", exception);
-			}
-		}
-
-	}
-
-	private static void runTestCaseSingleTransaction(String shaclPath, String dataPath, ExpectedResult expectedResult)
-		throws Exception
-	{
-
-		if (!dataPath.endsWith("/")) {
-			dataPath = dataPath + "/";
-		}
-
-		if (!shaclPath.endsWith("/")) {
-			shaclPath = shaclPath + "/";
-		}
-
-		ShaclSail shaclSail = new ShaclSail(new MemoryStore());
-		SailRepository shaclRepository = new SailRepository(shaclSail);
-		shaclRepository.initialize();
-		Utils.loadShapeData(shaclRepository, shaclPath + "shacl.ttl");
-
-		boolean exception = false;
-		boolean ran = false;
-
-		try (SailRepositoryConnection shaclSailConnection = shaclRepository.getConnection()) {
-			shaclSailConnection.begin(IsolationLevels.SNAPSHOT);
-
-			for (int j = 0; j < 100; j++) {
-
-				String name = dataPath + "query" + j + ".rq";
-				InputStream resourceAsStream = ShaclTest.class.getClassLoader().getResourceAsStream(name);
-				if (resourceAsStream == null) {
-					continue;
-				}
-
-				ran = true;
-				System.out.println(name);
-
-				try {
-					String query = IOUtil.readString(resourceAsStream);
-					shaclSailConnection.prepareUpdate(query).execute();
-
-				} catch (IOException e) {
-					e.printStackTrace();
-				}
-			}
-
-			try {
-				shaclSailConnection.commit();
-
-			} catch (RepositoryException sailException) {
-				exception = true;
-				System.out.println(sailException.getMessage());
-			}
-		}
-		if (ran) {
-			if (expectedResult == ExpectedResult.valid) {
-				assertFalse(exception);
-			}
-			else {
-				assertTrue(exception);
-			}
-		}
-
-	}
-
-	enum ExpectedResult {
-		valid,
-		invalid
 	}
 
 }
